Vagrant: Add support for Vagrant install (`vagrant up`)
Cf README, this adds support to install and setup a dev instance through
Vagrant. The Vagrantfile at the root handles the vagrant commands like
`vagrant up`.

Provisioning is done by the shell script `vagrant-provisionning.sh`,
which is a light wrapper around the `create-dev-setup.sh` which does
most of the job using the standard process, but limiting possible
side-effect by ensuring the environment on which the install is
performed is clean.

Based on an Ubuntu precise 12.04 LTS.